refactor(ContinualStonesWinningCondition)
- ContinualStones 에게 연속된 돌의 수를 세기를 위임한다
