+++ src/main/kotlin/omok/Application.kt
@@ -4,7 +4,8 @@ import omok.controller.OmokController
 import omok.model.rule.ban.DoubleFourForbiddenPlace
 import omok.model.rule.ban.DoubleOpenThreeForbiddenPlace
 import omok.model.rule.ban.OverlineForbiddenPlace
-import omok.model.rule.winning.FiveStonesWinningCondition
+import omok.model.rule.winning.ContinualStones
+import omok.model.rule.winning.ContinualStonesWinningCondition
 import omok.view.InputView
 import omok.view.OutputView
 
@@ -12,7 +13,7 @@ fun main() {
     OmokController(
         InputView(),
         OutputView(),
-        winningCondition = FiveStonesWinningCondition(),
+        winningCondition = ContinualStonesWinningCondition(ContinualStones()),
         blackStoneForbiddenPlaces =
             listOf(
                 DoubleFourForbiddenPlace(),
